# documentation/swagger.yaml
...
swagger: "2.0"
info:
  description: |
    This is the API documentation of the Imatra ICT project **eKolikko**.
    You may find additional information in the Introduction section as well as the RESTFUL API endpoints described above.
    # Introduction
    This API is documented in OpenAPI format and is based on the web application called eKolikko. The application introduces an innovative way for the citizens of Imatra to do small tasks for the community and for each other and in return to get paid in virtual credits for the tasks performed. Then, in turn, they could use those credits to pay for different services that the City of Imatra manages and provides.
    This API documentation explains all the endpoints used to create the back-end of the eKolikko solution. You may find the code of eKolikko in the following link: [eKolikko source code](https://github.com/imatra/eKolikko).
    # Cross-Origin Resource Sharing
    This API features Cross-Origin Resource Sharing (CORS) implemented in compliance with  [W3C spec](https://www.w3.org/TR/cors/).
    And that allows cross-domain communication from the browser.
    All responses have a wildcard same-origin which makes them completely public and accessible to everyone, including any code on any site.
    # Authentication
    eKolikko offers an OAuth2 approach on authentication. It provides a secure way to authenticate and get the Token bearer to further use to communicate with the application.

...
  version: 1.0.0
...
  license:
...
host: etesti.imatra.fi
basePath: /api/emoney/
...
schemes:
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: "Standard Authorization header using the Bearer scheme. Example: \"bearer {token}\". This bearer token comes from the authentication of an user using the login endpoint."
paths:
...
  /job/accept-applicant:
    post:
      tags:
      - Jobs
      summary: Accept a job applicant
      security:
      - Bearer: []
      description: The endpoint used from an user to accept a job appicant for one of his/her jobs.
      operationId: acceptApplicant
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: The JobApply object that contains applicantId and the jobId.
        required: true
        schema:
          $ref: '#/definitions/JobApply'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/AcceptApplicantResponse'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
  /job/reject-applicant:
    post:
      tags:
      - Jobs
      summary: Reject a job applicant
      security:
      - Bearer: []
      description: Reject a specific applicant for a specific job.
      operationId: rejectApplicant
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: The applicant id and the job id for the request.
        required: true
        schema:
          $ref: '#/definitions/JobApply' 
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/RejectApplicantResponse'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
  /job/active-job:
    get:
      tags:
      - Jobs
      summary: Get active jobs
      description: The endpoint used from an user to get all current active jobs.
      operationId: getActiveJobs
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/JobListResult'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
  /job/expired-job:
    get:
      tags:
      - Jobs
      summary: Get expired jobs
      description: The endpoint used to get all expired jobs.
      operationId: getExpiredJobs
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/JobListResult'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
...
  /job/my-jobs:
    get:
      tags:
      - Jobs
      summary: Get all jobs for the authenticated user.
      security:
      - Bearer: []
      description: The endpoint used to get all jobs for the authenticated user.
      operationId: getMyJobs
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/JobListResult'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
  /user/auth:
    post:
      tags:
      - User
      summary: Authenticate User
      description: The API endpoint that is used to authenticate an user.
      operationId: authenticateUser
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: The user email and password.
        required: true
        schema:
          $ref: '#/definitions/UserAuth' 
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/AuthenticationResponse'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
        400:
          description: Invalid username/password supplied
...
  /user/changepassword:
     post:
      tags:
      - User
      summary: Change user password
      security:
      - Bearer: []
      description: The API endpoint that is used to change the password of the authenticated user.
      operationId: changePassword
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: The user old password and new password.
        required: true
        schema:
          $ref: '#/definitions/ChangeUserPassword'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/ChangePasswordResponse'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
  /user/email:
     post:
      tags:
      - User
      summary: Verify user by email
      security:
      - Bearer: []
      description: The API endpoint that is used to verify user by email in the database.
      operationId: verifyUserByEmail
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: The user email
        required: true
        schema:
          $ref: '#/definitions/UserEmail'
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/UserEmailResponse'
        201:
          description: Created
        401:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        403:
          description: The unauthorized response
          schema:
            $ref: '#/definitions/Unauthorized'
        404:
          description: Not found
...
definitions:
...
  Unauthorized:
    type: object
    properties:
      message:
        type: string
        example: "Unauthorized access!!"
      page:
        type: object
        properties: {}
        example: null
      responseStatus:
        type: boolean
        example: false
      result:
        type: object
        properties: {}
        example: null
...
  RejectApplicantResponse:
    type: object
    properties:
      message:
        type: string
        example: "John Doe has been rejected for the Job."
      page:
        type: object
        properties: {}
        example: null
      responseStatus:
        type: boolean
        example: true
      result:
        type: string
        example: null
...
  AcceptApplicantResponse:
    type: object
    properties:
      message:
        type: string
        example: "John Doe has been selected for the Job."
      page:
        type: object
        properties: {}
        example: null
      responseStatus:
        type: boolean
        example: true
      result:
        type: string
        example: null
...
  AuthenticationResponse:
    type: object
    properties:
      message:
        type: string
        example: null
      page:
        type: object
        properties: {}
        example: null
      responseStatus:
        type: boolean
        example: true
      result:
       $ref: '#/definitions/AuthResult'
  AuthResult:
    type: object
    properties:
      isAdmin:
        type: boolean
        example: false
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiJ9.eyJzdWIittiJ0ZXN0QHRlc3QuY29tIiwidXNlcklkIjo4MSwiZW1haWwiOiJ0ZXN0QHRlc3QuY29tIiwiaWF0IjoxNTYzODQyMDM2LCJleHAiOjE1NjM4ODQwMzZ9.Hx9YbLFF3vOPUaae9cTfVUQs5wed9NAVbGf0axxCi0
  UserEmailResponse:
    type: object
    properties:
      message:
        type: string
        example: null
      page:
        type: object
        properties: {}
        example: null
      responseStatus:
        type: boolean
        example: true
      result:
        type: boolean
        example: true
...
  ChangePasswordResponse:
    type: object
    properties:
      message:
        type: string
        example: "Password has been changed successfully."
      page:
        type: object
        properties: {}
        example: null
      responseStatus:
        type: boolean
        example: true
      result:
        type: string
        example: null
...
  JobListResult:
     type: object
     properties:
      result:
        $ref: '#/definitions/JobResult'
      message:
        type: string
        example: "Success"
      responseStatus:
        type: boolean
        example: true
      page:
        type: object
        example: null
...
  JobResult:
    type: array
    items:
      type: object
      properties:
        version:
          type: integer
          example: 1
        id:
          type: integer
          example: 82
        jobTitle:
          type: string
          example: Clean my garden
        description:
          type: string
          example: I need someone to clean my garden.
        noOfPeople:
          type: string
          example: 3
        dueDate:
          type: string
          format: date-time
          example: "2019-07-24T00:00:00.000+0000"
        dueTime: 
          type: string
          example: "03:19:00"
        postedDate:
          type: string
          format: date-time
          example: "2019-07-22T22:06:20.000+0000"
        credits:
          type: number
          format: double
          example: 12.0
        address1: 
          type: string
          example: Punkkerikatu 2
        address2:
          type: string
          example: null
        postCode:
          type: string
          example: "53850"
        phoneNumber:
          type: string
          example: 04171113110
        qrFileName:
          type: string
          example: Fcymsg609997.png
        jobPosterId:
          type: integer
          format: int32
          example: 81
        jobPosterName:
          type: string
          example: John Doe
        totalSelected: 
          type: integer
          format: int32
          example: 1
        lat:
          type: number
          format: double
          example: 21.922323
        lng:
          type: number
          format: double
          example: 55.123123
...
  JobApply:
    type: object
    required:
    - applicantId
    - jobId
    properties:
      applicantId:
        type: integer
        format: int64
        example: 5
        description: The Applicant Id representing the user that applied for the job.
      jobId:
        type: integer
        format: int64
        example: 5
        description: The Job id representing the job that the user has applied to.
...
  UserEmail:
    type: object
    required:
    - email
    properties:
      email:
        type: string
        description: The user email address.
        example: johndoe@gmail.com
  ChangeUserPassword:
    type: object
    required:
    - oldPassword
    - newPassword
    - confirmPassword
    properties:
      oldPassword:
        type: string
        example: theoldpassword123
        description: The old password in plain text.
      newPassword:
        type: string
        example: newpassword123
        description: The new password in plain text.
      confirmPassword:
        type: string
        example: newpassword123
        description: The new password confirmation in plain text.
  UserAuth:
    type: object
    required:
    - email
    - password
    properties:
      email:
        type: string
        example: johndoe@gmail.com
        description: The email of the user.
      password:
        type: string
        example: testpassword123
        description: The plain password of the user.
...